@@ -7,19 +7,20 @@
     "build": "vue-cli-service build"
   },
   "dependencies": {
-    "@vue/composition-api": "^0.6.5",
     "core-js": "^2.6.5",
-    "vue": "^2.6.10",
+    "vue": "^3.0.0-beta.1",
     "vue-router": "^3.0.3",
-    "vuex": "^3.0.1"
+    "vuex": "^4.0.0-alpha.1"
   },
   "devDependencies": {
     "@vue/cli-plugin-babel": "^3.12.0",
     "@vue/cli-service": "^3.12.0",
+    "@vue/compiler-sfc": "^3.0.0-beta.1",
     "node-sass": "^4.12.0",
     "pug": "^3.0.0",
     "pug-plain-loader": "^1.0.0",
     "sass-loader": "^8.0.0",
-    "vue-template-compiler": "^2.6.10"
+    "vue-cli-plugin-vue-next": "~0.1.3",
+    "vue-template-compiler": "^2.6.11"
   }
 }
