@@ -1,11 +1,8 @@
-import Vue from 'vue'
 import Vuex from 'vuex'
-
-Vue.use(Vuex)
 
 import ui from './modules/ui'
 
-export default new Vuex.Store({
+export default Vuex.createStore({
   state: {
 
   },
@@ -18,4 +15,4 @@ export default new Vuex.Store({
   modules: {
     ui
   }
-})
+});
